feat: Provide regular nix flake template
To start using a nix4dev in a project, one can use the regular nix flake
template for the initial setup. There is no longer the need to run the
`init` command from the nix4dev repo.

The previous approach could be problematic from security perspective,
because one needed to run a command (that was hard to inspect). The
template approach is much better in this regard, because the `nix flake
init` command just creates some files that can be inspected before
actually running anything.