feat: Do not automatically add nix4dev to inputs (#123)
This causes problems when making customized nix4dev based devshell. The
flake with customized devshell might want to have nix4dev in it's
inputs, so that it's version is locked. Then the flake in a repo using
the customized devshell has nix4dev in input twice: on `inputs.nix4dev`
and `inputs.<custom-devshell>.inputs.nix4dev`. It is then quite easy to
use different versions for both instances, which causes various
problems.