test: Do not write local file in tofu/terraform tests (#130)
There are sometimes problems with the local provider on MacOS. The tests
use the provider packaged in nixpkgs, but sometimes TF starts to
complain that the provider is incompatible. So lets get rid of it. For
the purpouses of the test it is ok to test that TF plan runs successfuly
even when no resources are defined.